+++ packages/clerk-js/src/core/resources/SignUp.ts
@@ -14,11 +14,17 @@ import type {
   PreparePhoneNumberVerificationParams,
   PrepareVerificationParams,
   PrepareWeb3WalletVerificationParams,
-  SetActiveNavigate,
   SignUpAuthenticateWithWeb3Params,
   SignUpCreateParams,
   SignUpField,
+  SignUpFutureCreateParams,
+  SignUpFutureEmailCodeVerifyParams,
+  SignUpFutureFinalizeParams,
+  SignUpFuturePasswordParams,
+  SignUpFuturePhoneCodeSendParams,
+  SignUpFuturePhoneCodeVerifyParams,
   SignUpFutureResource,
+  SignUpFutureSSOParams,
   SignUpIdentificationField,
   SignUpJSON,
   SignUpJSONSnapshot,
@@ -30,9 +36,11 @@ import type {
 } from '@clerk/types';
 
 import {
+  generateSignatureWithBase,
   generateSignatureWithCoinbaseWallet,
   generateSignatureWithMetamask,
   generateSignatureWithOKXWallet,
+  getBaseIdentifier,
   getCoinbaseWalletIdentifier,
   getMetamaskIdentifier,
   getOKXWalletIdentifier,
@@ -87,7 +95,7 @@ export class SignUp extends BaseResource implements SignUpResource {
    *
    * An instance of `SignUpFuture`, which has a different API than `SignUp`, intended to be used in custom flows.
    */
-  __internal_future: SignUpFuture | null = new SignUpFuture(this);
+  __internal_future: SignUpFuture = new SignUpFuture(this);
 
   /**
    * @internal Only used for internal purposes, and is not intended to be used directly.
@@ -272,6 +280,21 @@ export class SignUp extends BaseResource implements SignUpResource {
       generateSignature: generateSignatureWithCoinbaseWallet,
       unsafeMetadata: params?.unsafeMetadata,
       strategy: 'web3_coinbase_wallet_signature',
+      legalAccepted: params?.legalAccepted,
+    });
+  };
+
+  public authenticateWithBase = async (
+    params?: SignUpAuthenticateWithWeb3Params & {
+      legalAccepted?: boolean;
+    },
+  ): Promise<SignUpResource> => {
+    const identifier = await getBaseIdentifier();
+    return this.authenticateWithWeb3({
+      identifier,
+      generateSignature: generateSignatureWithBase,
+      unsafeMetadata: params?.unsafeMetadata,
+      strategy: 'web3_base_signature',
       legalAccepted: params?.legalAccepted,
     });
   };
@@ -476,6 +499,8 @@ class SignUpFuture implements SignUpFutureResource {
   verifications = {
     sendEmailCode: this.sendEmailCode.bind(this),
     verifyEmailCode: this.verifyEmailCode.bind(this),
+    sendPhoneCode: this.sendPhoneCode.bind(this),
+    verifyPhoneCode: this.verifyPhoneCode.bind(this),
   };
 
   constructor(readonly resource: SignUp) {}
@@ -486,6 +511,27 @@ class SignUpFuture implements SignUpFutureResource {
 
   get unverifiedFields() {
     return this.resource.unverifiedFields;
+  }
+
+  get isTransferable() {
+    // TODO: we can likely remove the error code check as the status should be sufficient
+    return (
+      this.resource.verifications.externalAccount.status === 'transferable' &&
+      this.resource.verifications.externalAccount.error?.code === 'external_account_exists'
+    );
+  }
+
+  get existingSession() {
+    if (
+      (this.resource.verifications.externalAccount.status === 'failed' ||
+        this.resource.verifications.externalAccount.status === 'unverified') &&
+      this.resource.verifications.externalAccount.error?.code === 'identifier_already_signed_in' &&
+      this.resource.verifications.externalAccount.error?.meta?.sessionId
+    ) {
+      return { sessionId: this.resource.verifications.externalAccount.error?.meta?.sessionId };
+    }
+
+    return undefined;
   }
 
   private async getCaptchaToken(): Promise<{
@@ -503,7 +549,19 @@ class SignUpFuture implements SignUpFutureResource {
     return { captchaToken, captchaWidgetType, captchaError };
   }
 
-  async password({ emailAddress, password }: { emailAddress: string; password: string }): Promise<{ error: unknown }> {
+  async create(params: SignUpFutureCreateParams): Promise<{ error: unknown }> {
+    const { transfer } = params;
+    return runAsyncResourceTask(this.resource, async () => {
+      const { captchaToken, captchaWidgetType, captchaError } = await this.getCaptchaToken();
+      await this.resource.__internal_basePost({
+        path: this.resource.pathRoot,
+        body: { transfer, captchaToken, captchaWidgetType, captchaError },
+      });
+    });
+  }
+
+  async password(params: SignUpFuturePasswordParams): Promise<{ error: unknown }> {
+    const { emailAddress, password } = params;
     return runAsyncResourceTask(this.resource, async () => {
       const { captchaToken, captchaWidgetType, captchaError } = await this.getCaptchaToken();
 
@@ -530,7 +588,8 @@ class SignUpFuture implements SignUpFutureResource {
     });
   }
 
-  async verifyEmailCode({ code }: { code: string }): Promise<{ error: unknown }> {
+  async verifyEmailCode(params: SignUpFutureEmailCodeVerifyParams): Promise<{ error: unknown }> {
+    const { code } = params;
     return runAsyncResourceTask(this.resource, async () => {
       await this.resource.__internal_basePost({
         body: { strategy: 'email_code', code },
@@ -539,7 +598,62 @@ class SignUpFuture implements SignUpFutureResource {
     });
   }
 
-  async finalize({ navigate }: { navigate?: SetActiveNavigate }): Promise<{ error: unknown }> {
+  async sendPhoneCode(params: SignUpFuturePhoneCodeSendParams): Promise<{ error: unknown }> {
+    const { phoneNumber, channel = 'sms' } = params;
+    return runAsyncResourceTask(this.resource, async () => {
+      if (!this.resource.id) {
+        const { captchaToken, captchaWidgetType, captchaError } = await this.getCaptchaToken();
+        await this.resource.__internal_basePost({
+          path: this.resource.pathRoot,
+          body: { phoneNumber, captchaToken, captchaWidgetType, captchaError },
+        });
+      }
+
+      await this.resource.__internal_basePost({
+        body: { strategy: 'phone_code', channel },
+        action: 'prepare_verification',
+      });
+    });
+  }
+
+  async verifyPhoneCode(params: SignUpFuturePhoneCodeVerifyParams): Promise<{ error: unknown }> {
+    const { code } = params;
+    return runAsyncResourceTask(this.resource, async () => {
+      await this.resource.__internal_basePost({
+        body: { strategy: 'phone_code', code },
+        action: 'attempt_verification',
+      });
+    });
+  }
+
+  async sso(params: SignUpFutureSSOParams): Promise<{ error: unknown }> {
+    const { strategy, redirectUrl, redirectCallbackUrl } = params;
+    return runAsyncResourceTask(this.resource, async () => {
+      const { captchaToken, captchaWidgetType, captchaError } = await this.getCaptchaToken();
+      await this.resource.__internal_basePost({
+        path: this.resource.pathRoot,
+        body: {
+          strategy,
+          redirectUrl: SignUp.clerk.buildUrlWithAuth(redirectCallbackUrl),
+          redirectUrlComplete: redirectUrl,
+          captchaToken,
+          captchaWidgetType,
+          captchaError,
+        },
+      });
+
+      const { status, externalVerificationRedirectURL } = this.resource.verifications.externalAccount;
+
+      if (status === 'unverified' && !!externalVerificationRedirectURL) {
+        windowNavigate(externalVerificationRedirectURL);
+      } else {
+        clerkInvalidFAPIResponse(status, SignUp.fapiClient.buildEmailAddress('support'));
+      }
+    });
+  }
+
+  async finalize(params?: SignUpFutureFinalizeParams): Promise<{ error: unknown }> {
+    const { navigate } = params || {};
     return runAsyncResourceTask(this.resource, async () => {
       if (!this.resource.createdSessionId) {
         throw new Error('Cannot finalize sign-up without a created session.');