+++ packages/localizations/src/de-DE.ts
@@ -197,7 +197,7 @@ export const deDE: LocalizationResource = {
   footerPageLink__privacy: 'Privatsphäre',
   footerPageLink__terms: 'Bedingungen',
   formButtonPrimary: 'Fortsetzen',
-  formButtonPrimary__verify: 'Verify',
+  formButtonPrimary__verify: 'Verifizieren',
   formFieldAction__forgotPassword: 'Passwort vergessen?',
   formFieldError__matchingPasswords: 'Passwörter stimmen überein.',
   formFieldError__notMatchingPasswords: 'Passwörter stimmen nicht überein.',
@@ -227,7 +227,7 @@ export const deDE: LocalizationResource = {
   formFieldLabel__apiKeyExpiration: 'Ablaufdatum',
   formFieldLabel__apiKeyName: 'Name',
   formFieldLabel__automaticInvitations: 'Aktivieren Sie automatische Einladungen für diese Domain',
-  formFieldLabel__backupCode: 'Sicherungscode',
+  formFieldLabel__backupCode: 'Wiederherstellungscode',
   formFieldLabel__confirmDeletion: 'Bestätigung',
   formFieldLabel__confirmPassword: 'Passwort bestätigen',
   formFieldLabel__currentPassword: 'Aktuelles Passwort',
@@ -509,7 +509,7 @@ export const deDE: LocalizationResource = {
     alternativeMethods: {
       actionLink: 'Klicken Sie hier, um eine alternative Methode zu verwenden',
       actionText: 'Verwenden Sie eine alternative Verifizierungsmethode',
-      blockButton__backupCode: 'Mit Backup-Code verifizieren',
+      blockButton__backupCode: 'Mit Wiederherstellungscode verifizieren',
       blockButton__emailCode: 'Mit E-Mail-Code verifizieren',
       blockButton__passkey: 'Verwenden Sie Ihren Passkey',
       blockButton__password: 'Mit Passwort verifizieren',
@@ -524,8 +524,9 @@ export const deDE: LocalizationResource = {
       title: 'Verifizierung erforderlich',
     },
     backupCodeMfa: {
-      subtitle: 'Verwenden Sie den Backup-Code, der Ihnen bei der Registrierung zur Verfügung gestellt wurde.',
-      title: 'Backup-Code Verifizierung',
+      subtitle:
+        'Verwenden Sie den Wiederherstellungscode, der Ihnen bei der Registrierung zur Verfügung gestellt wurde.',
+      title: 'Wiederherstellungscode Verifizierung',
     },
     emailCode: {
       formTitle: 'Geben Sie den Code ein, den wir an Ihre E-Mail-Adresse gesendet haben.',
@@ -577,7 +578,7 @@ export const deDE: LocalizationResource = {
     alternativeMethods: {
       actionLink: 'Hilfe',
       actionText: 'Haben Sie keine davon?',
-      blockButton__backupCode: 'Verwenden Sie einen Backup-Code',
+      blockButton__backupCode: 'Verwenden Sie einen Wiederherstellungscode',
       blockButton__emailCode: 'Code an {{identifier}} senden',
       blockButton__emailLink: 'Link senden an {{identifier}}',
       blockButton__passkey: 'Melden Sie sich mit Ihrem Passkey an',
@@ -601,7 +602,7 @@ export const deDE: LocalizationResource = {
     },
     backupCodeMfa: {
       subtitle: 'weiter zu {{applicationName}}',
-      title: 'Geben Sie einen Backup-Code ein',
+      title: 'Geben Sie einen Wiederherstellungscode ein',
     },
     emailCode: {
       formTitle: 'Bestätigungscode',
@@ -962,16 +963,16 @@ export const deDE: LocalizationResource = {
       actionLabel__copy: 'Kopiere alle',
       actionLabel__download: 'Laden Sie .txt herunter',
       actionLabel__print: 'Drucken',
-      infoText1: 'Backup-Codes werden für dieses Konto aktiviert.',
+      infoText1: 'Wiederherstellungscodes werden für dieses Konto aktiviert.',
       infoText2:
-        'Halten Sie die Backup-Codes geheim und bewahren Sie sie sicher auf. Sie können Sicherungscodes neu generieren, wenn Sie vermuten, dass sie kompromittiert wurden.',
+        'Halten Sie die Wiederherstellungscodes geheim und bewahren Sie sie sicher auf. Sie können Wiederherstellungscodes neu generieren, wenn Sie vermuten, dass sie kompromittiert wurden.',
       subtitle__codelist: 'Bewahren Sie die Codes sicher auf und halten Sie sie geheim.',
       successMessage:
-        'Sicherungscodes sind jetzt aktiviert. Sie können eines davon verwenden, um sich bei Ihrem Konto anzumelden, wenn Sie den Zugriff auf Ihr Authentifizierungsgerät verlieren. Jeder Code kann nur einmal verwendet werden.',
+        'Wiederherstellungscodes sind jetzt aktiviert. Sie können eines davon verwenden, um sich bei Ihrem Konto anzumelden, wenn Sie den Zugriff auf Ihr Authentifizierungsgerät verlieren. Jeder Code kann nur einmal verwendet werden.',
       successSubtitle:
         'Sie können diese Codes verwenden, um sich bei Ihrem Konto anzumelden, wenn Sie den Zugriff auf Ihr Authentifizierungsgerät verlieren.',
-      title: 'Backup-Code-Verifizierung hinzufügen',
-      title__codelist: 'Sicherungscodes',
+      title: 'Wiederherstellungscode-Verifizierung hinzufügen',
+      title__codelist: 'Wiederherstellungscodes',
     },
     billingPage: {
       paymentHistorySection: {
@@ -1082,7 +1083,7 @@ export const deDE: LocalizationResource = {
       title: 'E-Mail-Adresse hinzufügen',
       verifyTitle: 'E-Mail Adresse verifizieren',
     },
-    formButtonPrimary__add: 'Add',
+    formButtonPrimary__add: 'Hinzufügen',
     formButtonPrimary__continue: 'Fortsetzen',
     formButtonPrimary__finish: 'Fertig',
     formButtonPrimary__remove: 'Entfernen',
@@ -1097,7 +1098,8 @@ export const deDE: LocalizationResource = {
       primaryButton__addPhoneNumber: 'Fügen Sie eine Telefonnummer hinzu',
       removeResource: {
         messageLine1: '{{identifier}} erhält bei der Anmeldung keine Bestätigungscodes mehr.',
-        messageLine2: 'Ihr Konto ist möglicherweise nicht so sicher. Bist du dir sicher, dass du weitermachen willst?',
+        messageLine2:
+          'Ihr Konto ist möglicherweise nicht so sicher. Sind Sie sich sicher, dass Sie fortfahren möchten?',
         successMessage: 'SMS-Code-Bestätigung in zwei Schritten wurde für {{mfaPhoneCode}} entfernt',
         title: 'Entfernen Sie die Bestätigung in zwei Schritten',
       },
@@ -1106,10 +1108,10 @@ export const deDE: LocalizationResource = {
       subtitle__unavailablePhoneNumbers:
         'Es sind keine Telefonnummern verfügbar, um sich für die SMS-Code-Bestätigung in zwei Schritten zu registrieren.',
       successMessage1:
-        'When signing in, you will need to enter a verification code sent to this phone number as an additional step.',
+        'Wenn Sie sich anmelden, müssen Sie einen Bestätigungscode eingeben, der an diese Telefonnummer gesendet wird.',
       successMessage2:
-        'Save these backup codes and store them somewhere safe. If you lose access to your authentication device, you can use backup codes to sign in.',
-      successTitle: 'SMS code verification enabled',
+        'Speichern Sie diese Wiederherstellungscodes und bewahren Sie sie an einem sicheren Ort auf. Falls Sie den Zugriff auf Ihr Authentifizierungsgerät verlieren, können Sie sich mit diesen Codes anmelden.',
+      successTitle: 'SMS-Code-Bestätigung aktiviert',
       title: 'SMS-Code-Bestätigung hinzufügen',
     },
     mfaTOTPPage: {
@@ -1176,8 +1178,8 @@ export const deDE: LocalizationResource = {
       },
       successMessage: '{{identifier}} wurde Ihrem Konto hinzugefügt.',
       title: 'Telefonnummer hinzufügen',
-      verifySubtitle: 'Enter the verification code sent to {{identifier}}',
-      verifyTitle: 'Verify phone number',
+      verifySubtitle: 'Geben Sie den Bestätigungscode ein, der an {{identifier}} gesendet wurde',
+      verifyTitle: 'Telefonnummer verifizieren',
     },
     plansPage: {
       title: 'Pläne',
@@ -1187,7 +1189,8 @@ export const deDE: LocalizationResource = {
       imageFormDestructiveActionSubtitle: 'Bild entfernen',
       imageFormSubtitle: 'Bild hochladen',
       imageFormTitle: 'Profilbild',
-      readonly: 'Your profile information has been provided by the enterprise connection and cannot be edited.',
+      readonly:
+        'Ihre Profilinformationen wurden durch Ihr Unternehmen bereitgestellt und können nicht bearbeitet werden.',
       successMessage: 'Ihr Profil wurde aktualisiert.',
       title: 'Profil aktualisieren',
     },
@@ -1203,7 +1206,7 @@ export const deDE: LocalizationResource = {
         primaryButton: 'Konto verbinden',
         subtitle__disconnected: 'Ihr Konto ist derzeit getrennt. Bitte verbinden Sie es erneut.',
         subtitle__reauthorize:
-          'The required scopes have been updated, and you may be experiencing limited functionality. Please re-authorize this application to avoid any issues',
+          'Die erforderlichen Berechtigungen wurden aktualisiert, wodurch Ihre Funktionalität möglicherweise eingeschränkt ist. Bitte autorisieren Sie diese Anwendung erneut, um Probleme zu vermeiden.',
         title: 'Verbundene Konten',
       },
       dangerSection: {
@@ -1226,10 +1229,10 @@ export const deDE: LocalizationResource = {
       mfaSection: {
         backupCodes: {
           actionLabel__regenerate: 'Codes neu generieren',
-          headerTitle: 'Backup-Codes',
+          headerTitle: 'Wiederherstellungscodes',
           subtitle__regenerate:
-            'Generieren Sie einen neuen Satz sicherer Backup-Codes. Alte Backup-Code werden gelöscht und können nicht mehr verwendet werden.',
-          title__regenerate: 'Backup-Codes neu generieren',
+            'Generieren Sie einen neuen Satz sicherer Wiederherstellungscodes. Alte Wiederherstellungscodes werden gelöscht und können nicht mehr verwendet werden.',
+          title__regenerate: 'Wiederherstellungscodes neu generieren',
         },
         phoneCode: {
           actionLabel__setDefault: 'Als Standard einstellen',